+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+* Updated toggle buttons https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/38
+* Fixed error where users could not upload the same file back to back https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/39
 * Issue #43 https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/43
 * Added Terminology API and mockked for tests
 * Updated toggle buttons https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/38